+++ index.js
@@ -1,6 +1,7 @@
 var _ = require("./lib")
 var parse = require("./parse")
 var stringify = require("./stringify")
+var canonicalize = stringify.canonicalize
 module.exports = Hugml
 
 function Hugml(namespaces) {
@@ -21,5 +22,10 @@ Hugml.prototype.stringify = function(obj) {
 	return stringify(this.aliases, obj)
 }
 
+Hugml.prototype.canonicalize = function(obj, path) {
+	return canonicalize(this.aliases, obj, path)
+}
+
 Hugml.parse = Hugml.prototype.parse.bind(Hugml.prototype)
 Hugml.stringify = Hugml.prototype.stringify.bind(Hugml.prototype)
+Hugml.canonicalize = Hugml.prototype.canonicalize.bind(Hugml.prototype)